Refactor I/O
