+++ tests/conftest.py
@@ -4,7 +4,7 @@ from typing import List, Union
 import numpy as np
 import pytest
 
-import btrack.dataio
+import btrack
 
 from ._utils import (
     RANDOM_SEED,
@@ -28,14 +28,14 @@ def test_real_objects():
     """
     Create a list of objects from real data.
     """
-    return btrack.dataio.import_CSV(TEST_DATA_PATH / "test_data.csv")
+    return btrack.io.import_CSV(TEST_DATA_PATH / "test_data.csv")
 
 
 def write_h5_file(file_path: os.PathLike, test_objects) -> os.PathLike:
     """
     Write a h5 file with test objects and return path.
     """
-    with btrack.dataio.HDF5FileHandler(file_path, "w") as h:
+    with btrack.io.HDF5FileHandler(file_path, "w") as h:
         h.write_objects(test_objects)
 
     return file_path
